+++ src/components/header/bottom-header/left-menu/left-menu.tsx
@@ -1,13 +1,30 @@
-import styles from '../bottom-header.module.css';
+import styles from './left-menu.module.css';
 import { NavLink } from 'react-router-dom';
 import { leftLinks } from './left-links-list';
 import classNames from 'classnames';
 import { SvgBuilder } from '../../../svg-builder/svg-builder';
 import { IconType } from '../../../svg-builder/enums';
 import { useState } from 'react';
+import { Button } from '../../../button/button';
+import { useEffect } from 'react';
 
 export const LeftNavMenu = () => {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
+
+  useEffect(() => {
+    const handleClickOutside = (event: Event) => {
+      if (
+        event.target instanceof Element &&
+        !event.target.closest(`.${styles.navigationLeftMenu}`) &&
+        !event.target.closest(`.${styles.burgerMenu}`)
+      ) {
+        setIsMenuOpen(false);
+      }
+    };
+
+    document.addEventListener('click', handleClickOutside);
+    return () => document.removeEventListener('click', handleClickOutside);
+  }, []);
 
   const openMenu = () => {
     setIsMenuOpen(!isMenuOpen);
@@ -15,6 +32,12 @@ export const LeftNavMenu = () => {
 
   return (
     <>
+      {isMenuOpen && (
+        <div className={styles.overlay} onClick={() => setIsMenuOpen(false)} />
+      )}
+      <Button className={styles.burgerMenu} onClick={openMenu}>
+        <SvgBuilder iconType={IconType.Burger} className={styles.burger} />
+      </Button>
       <nav
         className={classNames(styles.navigationLeftMenu, {
           [styles.active]: isMenuOpen,
@@ -33,11 +56,6 @@ export const LeftNavMenu = () => {
           );
         })}
       </nav>
-      <SvgBuilder
-        iconType={IconType.Burger}
-        className={styles.burgerMenu}
-        onClick={openMenu}
-      />
     </>
   );
 };
