+++ src/global.css
@@ -47,7 +47,7 @@ a:active {
   --clr-bg-2: #f6f5ff;
   --clr-line: #000;
   --shadow-1: 0 10px 40px rgb(0 0 0 / 5%), 0 -10px 40px rgb(0 0 0 / 5%);
-  --shadow-2: rgb(99 99 99 / 20%) 0px 2px 8px 0px;
+  --shadow-2: 0px 5px 15.2px 0px #00000021;
   --shadow-3: 0 12px 24px rgb(0 0 0 / 20%);
 }
