+++ src/pages/catalog-page/catalog/catalog.module.css
@@ -4,7 +4,7 @@
   width: 100%;
   align-items: center;
   justify-content: center;
-  gap: 53px;
+  gap: 49px;
 }
 
 @media (width <= 1416px) {
