+++ src/pages/catalog-page/product-card/product-card.tsx
@@ -3,7 +3,7 @@ import { catalogStore } from '../../../store/catalog-store';
 import { useEffect } from 'react';
 import { messages } from './messages';
 import { observer } from 'mobx-react-lite';
-import DEFAULT_IMAGE from '../../../../assets/images/placeholder.jpg';
+import DEFAULT_IMAGE from '../../../../assets/images/placeholder.png';
 import { Link } from 'react-router';
 import { PagePath } from '../../../router/enums';
 
@@ -18,8 +18,7 @@ export const ProductList = observer(() => {
     <>
       {productList.map(product => (
         <li key={product.id}>
-          <Link to={PagePath.root} className={styles.cardContainer}>
-            <span className={styles.name}>{product.name}</span>
+          <Link to={PagePath.productPage} className={styles.cardContainer}>
             <img
               className={styles.image}
               src={product.image}
@@ -28,6 +27,8 @@ export const ProductList = observer(() => {
                 event.currentTarget.src = DEFAULT_IMAGE;
               }}
             />
+            <span className={styles.name}>{product.name}</span>
+
             <p className={styles.description}>{product.description}</p>
             <div className={styles.priceContainer}>
               <p className={styles.price}>
